src/api/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
sqlalchemy==2.0.37
psycopg2-binary==2.9.10
pydantic==2.10.6
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyjwt==2.10.1
datetime==5.5
pytest==8.3.4
requests==2.32.3
supabase==2.12.0